add test for relationship between alt, az and RA, Dec
The old test catalogs had an incorrect altitude and azimuth
in the header (on the order of half an arcsecond).  I was
unable to diagnose where the error came from.  Even generating
the catalogs with 6 month old versions of the LSST Sims stack
failed to reproduce the bug.  I have updated the catalogs and
am adding this test to make sure that sort of error does
not slip by us again.