Remove the get_amp_image test
This test relies on function that use the old AFW amp API. We are now accesing
this information directly from obs_lsst and this test is no longer needed.  Also
remove those functions the test calls. They are no longer allowed as the
new Amplifier objects are immutable and this code was changing them.