Manually insert YAML elements.
It seems appending YAML elements in . notation doesn't work.
So, for now, manually insert the elements in the yaml sections.

This is dangerous! Hopefully this will be replaced by a nested template
soon.  We have to be careful to not let config files get out of sync.
Now nothing should be using the core imsim-config.yaml file.

Also, update two tests to use the instcat config file.